HDX-9781 Move population view as table
