+++ src/hapi_schema/db_population.py
@@ -3,7 +3,6 @@
 from datetime import datetime
 
 from sqlalchemy import (
-    Boolean,
     DateTime,
     Enum,
     ForeignKey,
@@ -96,30 +95,3 @@ view_params_population = ViewParams(
         )
     ),
 )
-
-
-class DBPopulationVAT(Base):
-    __tablename__ = "population_vat"
-    resource_hdx_id: Mapped[str] = mapped_column(String(36))
-    admin2_ref: Mapped[int] = mapped_column(Integer, primary_key=True)
-    gender: Mapped[str] = mapped_column(String(11), primary_key=True)
-    age_range: Mapped[str] = mapped_column(String(32), primary_key=True)
-    min_age: Mapped[int] = mapped_column(Integer, index=True)
-    max_age: Mapped[int] = mapped_column(Integer, index=True)
-    population: Mapped[int] = mapped_column(Integer)
-    reference_period_start: Mapped[datetime] = mapped_column(
-        DateTime, index=True
-    )
-    reference_period_end: Mapped[datetime] = mapped_column(
-        DateTime, index=True
-    )
-    location_code: Mapped[str] = mapped_column(String(128))
-    location_name: Mapped[str] = mapped_column(String(512))
-    admin1_code: Mapped[str] = mapped_column(String(128))
-    admin1_name: Mapped[str] = mapped_column(String(512))
-    admin1_is_unspecified: Mapped[bool] = mapped_column(Boolean)
-    location_ref: Mapped[int] = mapped_column(Integer)
-    admin2_code: Mapped[str] = mapped_column(String(128))
-    admin2_name: Mapped[str] = mapped_column(String(512))
-    admin2_is_unspecified: Mapped[bool] = mapped_column(Boolean)
-    admin1_ref: Mapped[int] = mapped_column(Integer)
