+++ tests/test_population.py
@@ -23,31 +23,6 @@ def test_population_view(run_view_test):
     )
 
 
-def test_population_vat(
-    run_indexes_test, run_columns_test, run_primary_keys_test
-):
-    """Check that the population view as table is correct - columns match, expected indexes present"""
-    expected_primary_keys = [
-        "admin2_ref",
-        "gender",
-        "age_range",
-    ]
-
-    expected_indexes = [
-        "min_age",
-        "max_age",
-        "reference_period_start",
-        "reference_period_end",
-    ]
-    run_columns_test(
-        "population_vat",
-        "population_view",
-        view_params_population,
-    )
-    run_indexes_test("population_vat", expected_indexes)
-    run_primary_keys_test("population_vat", expected_primary_keys)
-
-
 @pytest.fixture
 def base_parameters():
     return dict(
